refactor: fix naming related to carnet section item
